Refactor/#8 to single module (#9)
* [REFACT] .gradle 파일 수정

* [REFACT] 코어 및 도메인 모듈 제거
