+++ build.gradle
@@ -1,39 +1,54 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '3.2.2' apply(false)
-    id 'io.spring.dependency-management' version '1.1.4' apply(false)
+    id 'org.springframework.boot' version '3.2.2'
+    id 'io.spring.dependency-management' version '1.1.4'
 }
 
-allprojects {
-    group = 'io.oeid'
-    version = '0.0.1-SNAPSHOT'
-    java {
-        sourceCompatibility = 17
-        targetCompatibility = 17
-    }
-    repositories {
-        mavenCentral()
-    }
+group = 'io.oeid'
+version = '0.0.1-SNAPSHOT'
+java {
+    sourceCompatibility = 17
+    targetCompatibility = 17
 }
 
-subprojects {
-    apply { plugin('java') }
-    apply { plugin('org.springframework.boot') }
-    apply { plugin('io.spring.dependency-management') }
+repositories {
+    mavenCentral()
+}
 
-    configurations {
-        compileOnly {
-            extendsFrom annotationProcessor
-        }
-    }
+dependencies {
+    // Spring Boot Web
+    implementation 'org.springframework.boot:spring-boot-starter-web:3.2.2'
+    // Spring Boot Security
+    implementation 'org.springframework.boot:spring-boot-starter-security:3.2.2'
+    // Spring Boot Data JPA
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.2.2'
+    // Lombok
+    compileOnly 'org.projectlombok:lombok:1.18.30'
+    annotationProcessor 'org.projectlombok:lombok:1.18.30'
+    // QueryDSL
+    implementation 'com.querydsl:querydsl-jpa:5.1.0:jakarta'
+    annotationProcessor 'com.querydsl:querydsl-apt:5.1.0:jakarta'
+    annotationProcessor "jakarta.annotation:jakarta.annotation-api:2.1.1"
+    annotationProcessor "jakarta.persistence:jakarta.persistence-api:3.1.0"
+    // MySQL
+    runtimeOnly 'com.mysql:mysql-connector-j:8.2.0'
+    // Spring Boot Security
+    implementation 'org.springframework.boot:spring-boot-starter-security:3.2.2'
+    // Spring Boot OAuth2 Client
+    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client:3.2.2'
+    // JWT
+    implementation 'com.auth0:java-jwt:4.4.0'
 
-    dependencies {
-        // Lombok
-        compileOnly 'org.projectlombok:lombok:1.18.30'
-        annotationProcessor 'org.projectlombok:lombok:1.18.30'
-    }
+    // SpringBoot Test
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:3.2.2'
+    // TestContainers
+    testImplementation 'org.springframework.boot:spring-boot-testcontainers:3.2.2'
+    testImplementation 'org.testcontainers:junit-jupiter:1.19.3'
+    testImplementation 'org.testcontainers:mysql:1.19.3'
+    // Spring Security Test
+    testImplementation 'org.springframework.security:spring-security-test:6.1.3'
+}
 
-    test {
-        useJUnitPlatform()
-    }
+test {
+    useJUnitPlatform()
 }
