--- core/core-jpa/build.gradle
@@ -1,19 +1,0 @@
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
-
-dependencies {
-    // SpringBoot Data JPA
-    implementation 'org.springframework.boot:spring-boot-starter-web:3.2.2'
-    // QueryDSL
-    implementation 'com.querydsl:querydsl-jpa:5.1.0:jakarta'
-    annotationProcessor 'com.querydsl:querydsl-apt:5.1.0:jakarta'
-    annotationProcessor "jakarta.annotation:jakarta.annotation-api:2.1.1"
-    annotationProcessor "jakarta.persistence:jakarta.persistence-api:3.1.0"
-    // MySQL
-    runtimeOnly 'com.mysql:mysql-connector-j:8.2.0'
-}
