--- core/core-jpa/src/main/java/io/oeid/core_jpa/configuration/QueryDSLConfiguration.java
@@ -1,19 +1,0 @@
-package io.oeid.core_jpa.configuration;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class QueryDSLConfiguration {
-
-    private final EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
