--- core/core-security/build.gradle
@@ -1,21 +1,0 @@
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
-
-dependencies {
-    implementation project(':domain:domain-user')
-    // Spring Boot Web
-    implementation 'org.springframework.boot:spring-boot-starter-web:3.2.2'
-    // Spring Boot Security
-    implementation 'org.springframework.boot:spring-boot-starter-security:3.2.2'
-    // Spring Boot OAuth2 Client
-    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client:3.2.2'
-    // JWT
-    implementation 'com.auth0:java-jwt:4.4.0'
-    // Spring Boot Data JPA
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.2.2'
-}
