--- core/core-security/src/main/java/io/oeid/core_security/CoreSecurityApplication.java
@@ -1,9 +1,0 @@
-package io.oeid.core_security;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages = {"io.oeid"})
-// TODO: 추후 API 모듈 하나라도 생성하는 경우에 제거
-public class CoreSecurityApplication {
-
-}
