--- core/core-security/src/main/java/io/oeid/core_security/jwt/JwtAccessDeniedHandler.java
@@ -1,28 +1,0 @@
-package io.oeid.core_security.jwt;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-
-    private final HandlerExceptionResolver resolver;
-
-    public JwtAccessDeniedHandler(
-        @Qualifier("handlerExceptionResolver") HandlerExceptionResolver resolver) {
-        this.resolver = resolver;
-    }
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response,
-        AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        resolver.resolveException(request, response, null, accessDeniedException);
-    }
-}
