--- core/core-security/src/main/java/io/oeid/core_security/jwt/JwtAuthenticationToken.java
@@ -1,47 +1,0 @@
-package io.oeid.core_security.jwt;
-
-import java.util.Collection;
-import lombok.Getter;
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-
-@Getter
-public class JwtAuthenticationToken extends AbstractAuthenticationToken {
-
-    private final Object principal;
-    private Object credentials;
-
-    public JwtAuthenticationToken(Object principal, Object credentials) {
-        super(null);
-        this.principal = principal;
-        this.credentials = credentials;
-        this.setAuthenticated(false);
-    }
-
-    public JwtAuthenticationToken(Object principal, Object credentials,
-        Collection<? extends GrantedAuthority> authorities) {
-        super(authorities);
-        this.principal = principal;
-        this.credentials = credentials;
-        super.setAuthenticated(true);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof JwtAuthenticationToken token)) {
-            return false;
-        }
-        return token.principal == this.principal && token.credentials == this.credentials;
-    }
-
-    @Override
-    public int hashCode() {
-        int result = super.hashCode();
-        result = 31 * result + (principal != null ? principal.hashCode() : 0);
-        result = 31 * result + (credentials != null ? credentials.hashCode() : 0);
-        return result;
-    }
-}
