--- domain/build.gradle
@@ -1,7 +1,0 @@
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
