--- domain/domain-user/build.gradle
@@ -1,17 +1,0 @@
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
-
-dependencies {
-    implementation project(':core:core-jpa')
-    // Spring Boot Web
-    implementation 'org.springframework.boot:spring-boot-starter-web:3.2.2'
-    // Spring Boot Security
-    implementation 'org.springframework.boot:spring-boot-starter-security:3.2.2'
-    // Spring Boot Data JPA
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.2.2'
-}
