--- domain/domain-user/src/main/java/io/oeid/domain_user/domain/Role.java
@@ -1,5 +1,0 @@
-package io.oeid.domain_user.domain;
-
-public enum Role {
-    ROLE_USER
-}
