--- domain/domain-user/src/main/java/io/oeid/domain_user/infrastructure/UserJpaRepository.java
@@ -1,10 +1,0 @@
-package io.oeid.domain_user.infrastructure;
-
-import io.oeid.domain_user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserJpaRepository extends JpaRepository<User, Long> {
-
-}
