+++ src/main/java/io/oeid/mogakgo/MogakGoApplication.java
@@ -1,0 +1,13 @@
+package io.oeid.mogakgo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MogakGoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MogakGoApplication.class, args);
+    }
+
+}
