+++ src/main/java/io/oeid/mogakgo/core/configuration/SecurityConfiguration.java
@@ -1,0 +1,58 @@
+package io.oeid.mogakgo.core.configuration;
+
+
+import io.oeid.mogakgo.core.security.jwt.JwtAccessDeniedHandler;
+import io.oeid.mogakgo.core.security.jwt.JwtAuthenticationEntryPoint;
+import io.oeid.mogakgo.core.security.jwt.JwtAuthenticationFilter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+public class SecurityConfiguration {
+
+    private final OAuth2UserService<OAuth2UserRequest, OAuth2User> oAuth2UserService;
+    private final AuthenticationSuccessHandler successHandler;
+    private final JwtAuthenticationFilter jwtAuthenticationFilter;
+    private final JwtAccessDeniedHandler accessDeniedHandler;
+    private final JwtAuthenticationEntryPoint authenticationEntryPoint;
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
+        return httpSecurity
+            .httpBasic(AbstractHttpConfigurer::disable)
+            .csrf(AbstractHttpConfigurer::disable)
+            .formLogin(AbstractHttpConfigurer::disable)
+            .rememberMe(AbstractHttpConfigurer::disable)
+            .logout(AbstractHttpConfigurer::disable)
+            .headers(AbstractHttpConfigurer::disable)
+            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
+            .sessionManagement(
+                httpSecuritySessionManagementConfigurer -> httpSecuritySessionManagementConfigurer.sessionCreationPolicy(
+                    SessionCreationPolicy.STATELESS))
+            .oauth2Login(httpSecurityOAuth2LoginConfigurer -> {
+                httpSecurityOAuth2LoginConfigurer.userInfoEndpoint(
+                    userInfoEndpointConfig -> userInfoEndpointConfig.userService(oAuth2UserService)
+                );
+                httpSecurityOAuth2LoginConfigurer.successHandler(successHandler);
+            })
+            .exceptionHandling(httpSecurityExceptionHandlingConfigurer -> {
+                httpSecurityExceptionHandlingConfigurer.accessDeniedHandler(accessDeniedHandler);
+                httpSecurityExceptionHandlingConfigurer.authenticationEntryPoint(
+                    authenticationEntryPoint);
+            })
+            .build();
+    }
+}
