+++ src/main/java/io/oeid/mogakgo/domain/user/domain/enums/Role.java
@@ -1,0 +1,5 @@
+package io.oeid.mogakgo.domain.user.domain.enums;
+
+public enum Role {
+    ROLE_USER
+}
